feat: move put_events_with_filter_to_sqs fixture to conftest
